requirements.txt
azure-functions==1.5.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
idna==2.10
lxml==4.6.2
oauthlib==3.1.0
PySocks==1.7.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
tweepy==3.9.0
urllib3==1.26.2
